Revert "Progress on mini board for featured game"
This reverts commit fb19d4256ae1c1a179d74080d182a76b71f679a7.
